feat: add absolute paths for imports
