+++ src/index.tsx
@@ -8,14 +8,14 @@
 import { QueryClient, QueryClientProvider } from "react-query";
 
 //- Type Imports
-import { AppProps } from "./lib/types/app";
+import { AppProps } from "lib/types/app";
 
 //- Utils Imports
-import ChainGate from "./lib/util/ChainGate";
+import ChainGate from "lib/util/ChainGate";
 
 //- Provider Imports
-import Web3Provider from "./lib/providers/Web3Provider";
-import ZnsSdkProvider from "./lib/providers/ZnsSdkProvider";
+import Web3Provider from "lib/providers/Web3Provider";
+import ZnsSdkProvider from "lib/providers/ZnsSdkProvider";
 
 //- Container Imports
 import App from "./App";
