+++ src/lib/hooks/useWeb3.ts
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { Web3Context } from "../providers/Web3Provider";
+import { Web3Context } from "lib/providers/Web3Provider";
 
 export function useWeb3() {
   return useContext(Web3Context);
