+++ src/pages/ZNS/ZNS.tsx
@@ -1,25 +1,13 @@
-//- React Imports
-import { useState } from "react";
-
 //- Lib Imports
-import { useDomainMetrics } from "../../lib/hooks/useDomainMetrics";
-import { ethers } from "ethers";
+import { useDomainMetrics } from "lib/hooks/useDomainMetrics";
+import { formatEthers } from "lib/util/number";
 
 //- Style Imports
 import styles from "./ZNS.module.scss";
 
-export const formatNumber = (number: number | string) => {
-  return Number(number).toLocaleString();
-};
-
-export const formatEthers = (number: string) => {
-  const asNumber = Number(ethers.utils.formatEther(number));
-  return formatNumber(asNumber);
-};
-
 const ZNS = () => {
   const { tradeData } = useDomainMetrics(
-    "0x6ac9b49d77117984ae21556480c5585d36b2fb94545ab0adbc270a55bb59931b"
+    "0xd679f47d74935f037b9d4d507a17acdcffeb9672d82778e0adfd4bb583c5f2fc"
   );
 
   console.log(tradeData);
