+++ tsconfig.json
@@ -8,11 +8,7 @@
     "forceConsistentCasingInFileNames": true,
     "isolatedModules": true,
     "jsx": "react-jsx",
-    "lib": [
-      "dom",
-      "dom.iterable",
-      "esnext"
-    ],
+    "lib": ["dom", "dom.iterable", "esnext"],
     "module": "CommonJS",
     "moduleResolution": "node",
     "noFallthroughCasesInSwitch": true,
@@ -23,13 +19,11 @@
     "strict": true,
     "strictNullChecks": false,
     "target": "es5",
+    "paths": {
+      "src/*": ["./src/*"]
+    }
   },
-  "include": [
-    "src"
-  ],
-  "exclude": [
-    "src/**/*.test.*",
-    "src/**/*.test.d.*",
-    "src/setupTests.ts"
-  ]
+  "include": ["src"],
+
+  "exclude": ["src/**/*.test.*", "src/**/*.test.d.*", "src/setupTests.ts"]
 }
